Fix NPE in LocationService.  Implement new Logger module.  Migrate logging methods to new Logger module.  Implement new Logger.uploadLog method
