+++ example/ios/Flutter/flutter_export_environment.sh
@@ -6,6 +6,6 @@ export "FLUTTER_TARGET=/Users/chris/workspace/background-geolocation/flutter/flu
 export "FLUTTER_BUILD_DIR=build"
 export "SYMROOT=${SOURCE_ROOT}/../build/ios"
 export "FLUTTER_FRAMEWORK_DIR=/Users/chris/development/flutter/bin/cache/artifacts/engine/ios"
-export "FLUTTER_BUILD_NAME=1.2.52"
-export "FLUTTER_BUILD_NUMBER=9"
+export "FLUTTER_BUILD_NAME=1.3.01"
+export "FLUTTER_BUILD_NUMBER=10"
 export "TRACK_WIDGET_CREATION=true"
