+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: flutter_background_geolocation
 description: The most sophisticated background location-tracking & geofencing module with battery-conscious motion-detection intelligence for iOS and Android.
-version: 1.3.0
+version: 1.3.1
 author: Chris Scott, Transistor Software <info@transistorsoft.com>
 homepage: https://github.com/transistorsoft/flutter_background_geolocation
